+++ report-service/src/main/java/gov/cdc/dataingestion/config/DataSourceConfig.java
@@ -1,4 +1,4 @@
-package gov.cdc.dataingestion.report.config;
+package gov.cdc.dataingestion.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
