+++ report-service/src/main/java/gov/cdc/dataingestion/report/integration/conversion/interfaces/IHL7ToFHIRConversion.java
@@ -1,7 +1,7 @@
 package gov.cdc.dataingestion.report.integration.conversion.interfaces;
 
 import gov.cdc.dataingestion.report.repository.model.HL7toFhirModel;
-import gov.cdc.dataingestion.report.repository.model.ValidatedELRModel;
+import gov.cdc.dataingestion.validation.repository.model.ValidatedELRModel;
 
 public interface IHL7ToFHIRConversion {
     HL7toFhirModel ConvertHL7v2ToFhir(ValidatedELRModel validatedModel, String topicName);
