+++ report-service/src/main/java/gov/cdc/dataingestion/report/repository/model/HL7toFhirModel.java
@@ -1,9 +1,7 @@
 package gov.cdc.dataingestion.report.repository.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
+import org.hibernate.annotations.GenericGenerator;
 
 import java.sql.Timestamp;
 
@@ -11,14 +9,27 @@ import java.sql.Timestamp;
 @Table(name = "elr_fhir")
 public class HL7toFhirModel {
     @Id
+    @GenericGenerator(name = "generator", strategy = "guid", parameters = {})
+    @GeneratedValue(generator = "generator")
+    @Column(name = "id" , columnDefinition="uniqueidentifier")
     private String id;
-    private String raw_id;
+
+    @Column(name = "raw_id")
+    private String rawId;
     @Column(name = "fhir_msg")
     private String fhirMessage;
-    private String created_by;
-    private String updated_by;
-    private Timestamp created_on;
-    private Timestamp updated_on;
+    @Transient
+    @Column(name = "created_on")
+    private Timestamp CreatedOn;
+
+    @Column(name = "updated_on")
+    private Timestamp updatedOn;
+
+    @Column(name = "created_by")
+    private String createdBy;
+
+    @Column(name = "updated_by")
+    private String updatedBy;
 
 
     public String getId() {
@@ -29,12 +40,12 @@ public class HL7toFhirModel {
         this.id = id;
     }
 
-    public String getRaw_id() {
-        return raw_id;
+    public String getRawId() {
+        return rawId;
     }
 
-    public void setRaw_id(String raw_id) {
-        this.raw_id = raw_id;
+    public void setRawId(String rawId) {
+        this.rawId = rawId;
     }
 
     public String getFhirMessage() {
@@ -45,35 +56,35 @@ public class HL7toFhirModel {
         this.fhirMessage = fhirMessage;
     }
 
-    public String getCreated_by() {
-        return created_by;
+    public Timestamp getCreatedOn() {
+        return CreatedOn;
     }
 
-    public void setCreated_by(String created_by) {
-        this.created_by = created_by;
+    public void setCreatedOn(Timestamp createdOn) {
+        CreatedOn = createdOn;
     }
 
-    public String getUpdated_by() {
-        return updated_by;
+    public Timestamp getUpdatedOn() {
+        return updatedOn;
     }
 
-    public void setUpdated_by(String updated_by) {
-        this.updated_by = updated_by;
+    public void setUpdatedOn(Timestamp updatedOn) {
+        this.updatedOn = updatedOn;
     }
 
-    public Timestamp getCreated_on() {
-        return created_on;
+    public String getCreatedBy() {
+        return createdBy;
     }
 
-    public void setCreated_on(Timestamp created_on) {
-        this.created_on = created_on;
+    public void setCreatedBy(String createdBy) {
+        this.createdBy = createdBy;
     }
 
-    public Timestamp getUpdated_on() {
-        return updated_on;
+    public String getUpdatedBy() {
+        return updatedBy;
     }
 
-    public void setUpdated_on(Timestamp updated_on) {
-        this.updated_on = updated_on;
+    public void setUpdatedBy(String updatedBy) {
+        this.updatedBy = updatedBy;
     }
 }
