--- report-service/src/main/java/gov/cdc/dataingestion/validation/config/DataSourceConfig.java
@@ -1,43 +1,0 @@
-package gov.cdc.dataingestion.validation.config;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import javax.sql.DataSource;
-
-@Configuration
-public class DataSourceConfig {
-    private static Logger logger = LoggerFactory.getLogger(DataSourceConfig.class);
-
-    @Value("${spring.datasource.driverClassName}")
-    private String className;
-
-    @Value("${spring.datasource.url}")
-    private String dbUrl;
-
-    @Value("${spring.datasource.username}")
-    private String userName;
-
-    @Value("${spring.datasource.password}")
-    private String password;
-    @Bean()
-    public DataSource dataSource() {
-        String driverClassName = this.className;
-        String dbUrl = this.dbUrl;
-        String dbUserName = this.userName;
-        String dbUserPassword = this.password;
-
-        DataSourceBuilder dataSourceBuilder = DataSourceBuilder.create();
-
-        dataSourceBuilder.driverClassName(driverClassName);
-        dataSourceBuilder.url(dbUrl);
-        dataSourceBuilder.username(dbUserName);
-        dataSourceBuilder.password(dbUserPassword);
-
-        return dataSourceBuilder.build();
-    }
-}
