--- report-service/src/main/java/gov/cdc/dataingestion/validation/config/KafkaConfig.java
@@ -1,52 +1,0 @@
-package gov.cdc.dataingestion.validation.config;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.core.*;
-
-import java.util.HashMap;
-import java.util.Map;
-@Slf4j
-@EnableKafka
-@Configuration
-public class KafkaConfig {
-    @Value("${kafka.validation.consumer.group-id}")
-    private String groupId = "";
-
-    @Bean
-    public ConsumerFactory<String, String> consumerFactory(@Value("${kafka.bootstrap-servers}") final String bootstrapServers,
-                                                           @Value("${kafka.validation.consumer.maxPollIntervalMs}") final String maxPollIntervalMs) {
-        final Map<String, Object> config = new HashMap<>();
-        config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        config.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
-        config.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        config.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-        config.put(ConsumerConfig.MAX_POLL_INTERVAL_MS_CONFIG, maxPollIntervalMs);
-        return new DefaultKafkaConsumerFactory<>(config);
-    }
-
-    @Bean
-    public ProducerFactory<String, String> producerFactory(@Value("${kafka.bootstrap-servers}") final String bootstrapServers) {
-        final Map<String, Object> config = new HashMap<>();
-        config.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
-        config.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        config.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-        return new DefaultKafkaProducerFactory<>(config);
-    }
-
-    @Bean
-    public KafkaTemplate<String, String> kafkaTemplate(final ProducerFactory<String, String> producerFactory,
-                                                       final ConsumerFactory<String, String> consumerFactory) {
-        // set factory for both producer and consumer
-        KafkaTemplate template = new KafkaTemplate<>(producerFactory);
-        template.setConsumerFactory(consumerFactory);
-        return template;
-    }
-}
