+++ report-service/src/main/java/gov/cdc/dataingestion/validation/controller/ValidationController.java
@@ -1,9 +1,9 @@
 package gov.cdc.dataingestion.validation.controller;
 
+import gov.cdc.dataingestion.report.repository.IRawELRRepository;
+import gov.cdc.dataingestion.report.repository.model.RawERLModel;
 import gov.cdc.dataingestion.validation.integration.service.KafkaProducerService;
-import gov.cdc.dataingestion.validation.model.RawERLModel;
 import gov.cdc.dataingestion.validation.model.constant.KafkaHeaderValue;
-import gov.cdc.dataingestion.validation.repository.RawELRRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -12,9 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.beans.factory.annotation.Value;
 
-import java.sql.Timestamp;
-import java.util.UUID;
-
 @Controller
 public class ValidationController {
 
@@ -22,7 +19,7 @@ public class ValidationController {
     String topicName;
 
     @Autowired
-    private RawELRRepository rawELRRepository;
+    private IRawELRRepository rawELRRepository;
     KafkaProducerService kafkaProducerService;
     public ValidationController(KafkaProducerService kafkaProducerService) {
         this.kafkaProducerService = kafkaProducerService;
