+++ report-service/src/main/java/gov/cdc/dataingestion/validation/integration/validator/HL7v2Validator.java
@@ -5,12 +5,10 @@ import ca.uhn.hl7v2.HapiContext;
 import ca.uhn.hl7v2.model.Message;
 import ca.uhn.hl7v2.parser.PipeParser;
 import ca.uhn.hl7v2.validation.impl.ValidationContextFactory;
+import gov.cdc.dataingestion.report.repository.model.RawERLModel;
 import gov.cdc.dataingestion.validation.integration.validator.interfaces.IHL7v2Validator;
-import gov.cdc.dataingestion.validation.model.RawERLModel;
-import gov.cdc.dataingestion.validation.model.ValidatedELRModel;
+import gov.cdc.dataingestion.validation.repository.model.ValidatedELRModel;
 import gov.cdc.dataingestion.validation.model.enums.MessageType;
-
-import java.sql.Timestamp;
 
 public class HL7v2Validator implements IHL7v2Validator {
     private HapiContext context;
@@ -18,7 +16,7 @@ public class HL7v2Validator implements IHL7v2Validator {
         this.context = context;
     }
 
-    public ValidatedELRModel MessageValidation(String id,  RawERLModel rawERLModel, String topicName) throws HL7Exception {
+    public ValidatedELRModel MessageValidation(String id, RawERLModel rawERLModel, String topicName) throws HL7Exception {
         String replaceSpecialCharacters;
         if (rawERLModel.getPayload().contains("\n")) {
             replaceSpecialCharacters = rawERLModel.getPayload().replaceAll("\n","\r");
