+++ report-service/src/main/java/gov/cdc/dataingestion/validation/integration/validator/interfaces/ICsvValidator.java
@@ -1,6 +1,6 @@
 package gov.cdc.dataingestion.validation.integration.validator.interfaces;
 
-import gov.cdc.dataingestion.validation.model.ValidatedELRModel;
+import gov.cdc.dataingestion.validation.repository.model.ValidatedELRModel;
 import org.springframework.stereotype.Component;
 import java.io.Reader;
 import java.util.List;
