+++ report-service/src/main/java/gov/cdc/dataingestion/validation/integration/validator/interfaces/IHL7v2Validator.java
@@ -1,8 +1,8 @@
 package gov.cdc.dataingestion.validation.integration.validator.interfaces;
 
 import ca.uhn.hl7v2.HL7Exception;
-import gov.cdc.dataingestion.validation.model.RawERLModel;
-import gov.cdc.dataingestion.validation.model.ValidatedELRModel;
+import gov.cdc.dataingestion.report.repository.model.RawERLModel;
+import gov.cdc.dataingestion.validation.repository.model.ValidatedELRModel;
 import org.springframework.stereotype.Component;
 
 @Component
