--- report-service/src/main/java/gov/cdc/dataingestion/validation/model/RawERLModel.java
@@ -1,91 +1,0 @@
-package gov.cdc.dataingestion.validation.model;
-
-import jakarta.persistence.*;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.sql.Date;
-import java.sql.Timestamp;
-
-@Entity
-@Table(name = "elr_raw")
-public class RawERLModel {
-
-    @Id
-    @GenericGenerator(name = "generator", strategy = "guid", parameters = {})
-    @GeneratedValue(generator = "generator")
-    @Column(name = "id" , columnDefinition="uniqueidentifier")
-    private String id;
-
-    @Column(name = "message_type")
-    private String type;
-    private String payload;
-
-    @Transient
-    @Column(name = "created_on")
-    private Timestamp CreatedOn;
-
-    @Column(name = "updated_on")
-    private Timestamp updatedOn;
-
-    @Column(name = "created_by")
-    private String createdBy;
-
-    @Column(name = "updated_by")
-    private String updatedBy;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getPayload() {
-        return payload;
-    }
-
-    public void setPayload(String payload) {
-        this.payload = payload;
-    }
-
-    public Timestamp getCreatedOn() {
-        return CreatedOn;
-    }
-
-    public void setCreatedOn(Timestamp createdOn) {
-        CreatedOn = createdOn;
-    }
-
-    public Timestamp getUpdatedOn() {
-        return updatedOn;
-    }
-
-    public void setUpdatedOn(Timestamp updatedOn) {
-        this.updatedOn = updatedOn;
-    }
-
-    public String getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(String createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public String getUpdatedBy() {
-        return updatedBy;
-    }
-
-    public void setUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-    }
-}
