--- report-service/src/main/java/gov/cdc/dataingestion/validation/model/ValidatedELRModel.java
@@ -1,113 +1,0 @@
-package gov.cdc.dataingestion.validation.model;
-
-import jakarta.persistence.*;
-import org.hibernate.annotations.GenericGenerator;
-
-import java.sql.Timestamp;
-
-@Entity
-@Table(name = "elr_validated")
-public class ValidatedELRModel {
-
-    @Id
-    @GenericGenerator(name = "generator", strategy = "guid", parameters = {})
-    @GeneratedValue(generator = "generator")
-    @Column(name = "id" , columnDefinition="uniqueidentifier")
-    private String id;
-
-    private String rawId;
-
-    @Column(name = "validated_message")
-    private String rawMessage;
-
-    @Column(name = "message_type")
-    private String messageType;
-
-    @Column(name = "message_version")
-    private String messageVersion;
-
-    @Transient
-    @Column(name = "created_on")
-    private Timestamp createdOn;
-
-    @Column(name = "updated_on")
-    private Timestamp updatedOn;
-
-    @Column(name = "created_by")
-    private String createdBy;
-
-    @Column(name = "updated_by")
-    private String updatedBy;
-
-    public String getRawMessage() {
-        return rawMessage;
-    }
-
-    public void setRawMessage(String rawMessage) {
-        this.rawMessage = rawMessage;
-    }
-
-    public String getMessageType() {
-        return messageType;
-    }
-
-    public void setMessageType(String messageType) {
-        this.messageType = messageType;
-    }
-
-    public String getMessageVersion() {
-        return messageVersion;
-    }
-
-    public void setMessageVersion(String messageVersion) {
-        this.messageVersion = messageVersion;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getRawId() {
-        return rawId;
-    }
-
-    public void setRawId(String rawId) {
-        this.rawId = rawId;
-    }
-
-    public Timestamp getCreatedOn() {
-        return createdOn;
-    }
-
-    public void setCreatedOn(Timestamp createdOn) {
-        this.createdOn = createdOn;
-    }
-
-    public Timestamp getUpdatedOn() {
-        return updatedOn;
-    }
-
-    public void setUpdatedOn(Timestamp updatedOn) {
-        this.updatedOn = updatedOn;
-    }
-
-    public String getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(String createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public String getUpdatedBy() {
-        return updatedBy;
-    }
-
-    public void setUpdatedBy(String updatedBy) {
-        this.updatedBy = updatedBy;
-    }
-}
