+++ report-service/src/main/java/gov/cdc/dataingestion/validation/repository/model/ValidatedELRModel.java
@@ -1,15 +1,21 @@
-package gov.cdc.dataingestion.report.repository.model;
+package gov.cdc.dataingestion.validation.repository.model;
 
 import jakarta.persistence.*;
+import org.hibernate.annotations.GenericGenerator;
 
 import java.sql.Timestamp;
 
 @Entity
 @Table(name = "elr_validated")
 public class ValidatedELRModel {
+
     @Id
+    @GenericGenerator(name = "generator", strategy = "guid", parameters = {})
+    @GeneratedValue(generator = "generator")
+    @Column(name = "id" , columnDefinition="uniqueidentifier")
     private String id;
 
+    @Column(name = "raw_id")
     private String rawId;
 
     @Column(name = "validated_message")
@@ -21,10 +27,18 @@ public class ValidatedELRModel {
     @Column(name = "message_version")
     private String messageVersion;
 
-    private Timestamp created_on;
-    private Timestamp updated_on;
-    private String created_by;
-    private String updated_by;
+    @Transient
+    @Column(name = "created_on")
+    private Timestamp createdOn;
+
+    @Column(name = "updated_on")
+    private Timestamp updatedOn;
+
+    @Column(name = "created_by")
+    private String createdBy;
+
+    @Column(name = "updated_by")
+    private String updatedBy;
 
     public String getRawMessage() {
         return rawMessage;
@@ -66,35 +80,35 @@ public class ValidatedELRModel {
         this.rawId = rawId;
     }
 
-    public Timestamp getCreated_on() {
-        return created_on;
+    public Timestamp getCreatedOn() {
+        return createdOn;
     }
 
-    public void setCreated_on(Timestamp created_on) {
-        this.created_on = created_on;
+    public void setCreatedOn(Timestamp createdOn) {
+        this.createdOn = createdOn;
     }
 
-    public Timestamp getUpdated_on() {
-        return updated_on;
+    public Timestamp getUpdatedOn() {
+        return updatedOn;
     }
 
-    public void setUpdated_on(Timestamp updated_on) {
-        this.updated_on = updated_on;
+    public void setUpdatedOn(Timestamp updatedOn) {
+        this.updatedOn = updatedOn;
     }
 
-    public String getCreated_by() {
-        return created_by;
+    public String getCreatedBy() {
+        return createdBy;
     }
 
-    public void setCreated_by(String created_by) {
-        this.created_by = created_by;
+    public void setCreatedBy(String createdBy) {
+        this.createdBy = createdBy;
     }
 
-    public String getUpdated_by() {
-        return updated_by;
+    public String getUpdatedBy() {
+        return updatedBy;
     }
 
-    public void setUpdated_by(String updated_by) {
-        this.updated_by = updated_by;
+    public void setUpdatedBy(String updatedBy) {
+        this.updatedBy = updatedBy;
     }
 }
