+++ report-service/src/main/resources/application.properties
@@ -3,18 +3,14 @@ spring.data.mongodb.port=27017
 spring.data.mongodb.database=reportdb
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
 
-
-
-
-
 # KAFKA
 kafka.bootstrap-servers=localhost:9092
-kafka.validation.consumer.group-id=validation-group
-kafka.validation.consumer.maxPollIntervalMs=30000
+kafka.group-id=data-ingestion-group
+kafka.consumer.maxPollIntervalMs=30000
+
+
 kafka.validation.producer.topic=elr_validated
 kafka.raw.producer.topic=elr_raw
-kafka.fhir-conversion.consumer.group-id=fhir-conversion-group
-kafka.fhir-conversion.consumer.maxPollIntervalMs=30000
 kafka.fhir-conversion.producer.topic="fhir_converted"
 kafka.fhir-conversion.consumer.max-retry=1
